refactor: move everything in @/utils to @/lib/utils
