+++ src/components/MenuItem.astro
@@ -1,6 +1,6 @@
 ---
 import { Image } from "astro:assets"
-import { formatCurrency } from "@/utils"
+import { formatCurrency } from "@/lib/utils"
 import type { MenuItem } from "@/types"
 
 interface Props {
