+++ src/components/order/OrderItem.astro
@@ -1,6 +1,6 @@
 ---
 import type { OrderReturnType } from "@/database/queries"
-import { formatDate, formatCurrency } from "@/utils"
+import { formatDate, formatCurrency } from "@/lib/utils"
 
 interface Props {
 	order: OrderReturnType[0]
